Bugfix and removed log

+++ uploader/reuploader.py
@@ -26,54 +26,57 @@ for directory in os.listdir(basedir):
     print directory
     for element in os.listdir(basedir + '/' + directory):
         if ".xml" == element[-4:] and element.replace(".xml", "") and 'Stampa' in directory:
-            print directory
             done = False
-            while not done:
+            print element.replace(".xml", "")
+            while done == False:
                 try:
                     time.sleep(1)
-                    r = requests.get("https://archive.org/metadata/" +  element.replace(".xml", ""), timeout=2 )
+                    r = requests.get("https://archive.org/metadata/" +  element.replace(".xml", ""), timeout=2)
                     done = True
                 except:
-                    print "Failed request"
+                    done = False
+                    print("Failed request")
 
-            #if r.text != '{}':
-            #    print("skipping " + element.replace(".xml", ""))
-            #else:
-                #and 'TO01157391_TO0324_62137_000000' in element:
-                #print 'uploading'
+            if r.text != '{}':
+                print("skipping " + element.replace(".xml", ""))
 
+            else:
                 bookObj = {}
                 part_name = ""
-
                 filename = basedir + '/' + directory + '/' +  element
 
                 bookObj = ASTOteca.parseXML(filename)
 
-                print 'zip -r -j ' + bookObj['identifier'] + '_images.zip ' + basedir.replace(' ', '\ ') + directory.replace(' ', '\ ') +  '/' + element.replace(".xml", "") + '/Archivio/*'
-                zip_exit_code = os.system('zip -r -j ' + bookObj['identifier'] + '_images.zip ' + basedir.replace(' ', '\ ') + directory.replace(' ', '\ ') +  '/' + element.replace(".xml", "") + '/Archivio/*')
-                #sys.exit(0)
-                if zip_exit_code != 0:
+                print basedir + directory +  '/' + element.replace(".xml", "") + '/Archivio/'
+                if os.path.exists(basedir + directory +  '/' + element.replace(".xml", "") + '/Archivio/'):
+                    print "Compressing"
+                    zip_exit_code = os.system('zip -r -j ' + bookObj['identifier'] + '_images.zip ' + basedir.replace(' ', '\ ') + directory.replace(' ', '\ ') +  '/' + element.replace(".xml", "") + '/Archivio/*')
+                    #sys.exit(0)
+                    if zip_exit_code != 0:
+                        print "ERROR in creating the pdf. Exit code is: " + str(zip_exit_code)
+                        sys.exit(zip_exit_code)
 
-                    print "ERROR in creating the pdf. Exit code is: " + str(zip_exit_code)
-                    sys.exit(zip_exit_code)
+                    print "Uploading"
+                    #print bookObj
+                    del bookObj['toc']
+                    del bookObj['cover_img']
+                    code = -1
+                    while code != 200:
+                        try:
+                            #print bookObj
+                            r = upload(bookObj['identifier'],files=bookObj['identifier'] + '_images.zip', metadata=bookObj)
+                            code = r[0].status_code
+                        except:
 
-                print "Uploading"
-                #print bookObj
-                code = -1
-                while code != 200:
-                    try:
-                        r = upload(bookObj['identifier'],files=identifier + '_images.zip', metadata=bookObj)
-                        code = r[0].status_code
-                    except:
-                        print "Failed upload"
+                            print "Failed upload"
 
-                print "Done with code " + str(code)
-                bookObj['resCode'] = code
+                    print "Done with code " + str(code)
+                    bookObj['resCode'] = code
 
-                #shutil.copytree(filename.replace(".xml","") + '/Internet/', filename.replace(".xml","") + /bookObj['identifier'])
-                #print bookObj['identifier']
-                #sys.exit(0)
-                data.append(bookObj)
-                thisbookdf = pd.DataFrame(data)
-                bookdf = bookdf.append(thisbookdf, ignore_index=True)
-                bookdf.to_csv('uploading-debug-book.csv', index=False, encoding='utf-8')
+                    #shutil.copytree(filename.replace(".xml","") + '/Internet/', filename.replace(".xml","") + /bookObj['identifier'])
+                    #print bookObj['identifier']
+                    #sys.exit(0)
+                    data.append(bookObj)
+                    thisbookdf = pd.DataFrame(data)
+                    bookdf = bookdf.append(thisbookdf, ignore_index=True)
+                    bookdf.to_csv('uploading-debug-book.csv', index=False, encoding='utf-8')
